+++ src/components/style_modules/Navbar.module.scss
@@ -8,7 +8,7 @@
   border-bottom: 2px solid $black-bis;
   margin-top: 0 !important;
 
-  .navbutton {
+  button {
     border: 0;
     border-radius: $radius;
     background: none;
